Widths are floats, not ints

+++ src/Document/Dictionary/DictionaryValue/Array/CIDFontWidths.php
@@ -2,6 +2,7 @@
 
 namespace PrinsFrank\PdfParser\Document\Dictionary\DictionaryValue\Array;
 
+use Override;
 use PrinsFrank\PdfParser\Document\Dictionary\DictionaryValue\Array\Item\ConsecutiveCIDWidth;
 use PrinsFrank\PdfParser\Document\Dictionary\DictionaryValue\Array\Item\RangeCIDWidth;
 use PrinsFrank\PdfParser\Document\Dictionary\DictionaryValue\DictionaryValue;
@@ -17,7 +18,7 @@ class CIDFontWidths implements DictionaryValue {
         $this->widths = $widths;
     }
 
-    public function getWidthForCharacterCode(int $characterCode): ?int {
+    public function getWidthForCharacterCode(int $characterCode): ?float {
         foreach ($this->widths as $widthItem) {
             if (($widthForCharacterCode = $widthItem->getWidthForCharacterCode($characterCode)) !== null) {
                 return $widthForCharacterCode;
@@ -27,6 +28,7 @@ class CIDFontWidths implements DictionaryValue {
         return null;
     }
 
+    #[Override]
     public static function fromValue(string $valueString): ?self {
         if (preg_match_all('/(?<startingCID>[0-9]+)\s+(?<CIDS>[0-9]+\s*[0-9.]+|\[[0-9. ]+\])/', $valueString, $matches, PREG_SET_ORDER) <= 0) {
             return null;
