+++ src/Document/Dictionary/DictionaryValue/Array/Item/ConsecutiveCIDWidth.php
@@ -3,14 +3,14 @@
 namespace PrinsFrank\PdfParser\Document\Dictionary\DictionaryValue\Array\Item;
 
 class ConsecutiveCIDWidth {
-    /** @param list<int> $widths */
+    /** @param list<float> $widths */
     public function __construct(
         public readonly int $cidStart,
         public readonly array $widths,
     ) {
     }
 
-    public function getWidthForCharacterCode(int $characterCode): ?int {
+    public function getWidthForCharacterCode(int $characterCode): ?float {
         return $this->widths[$characterCode - $this->cidStart] ?? null;
     }
 }
